requirements.txt
# Market data
yfinance>=0.2.31

# Data processing & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
ta>=0.11.0

# News & sentiment
feedparser>=6.0.0
requests>=2.31.0
transformers>=4.35.0
torch>=2.1.0

# Storage
# (sqlite3 is in the Python standard library)

# Scheduling
schedule>=1.2.0

# CLI
click>=8.1.0

# Configuration
python-dotenv>=1.0.0
